+++ halo2_proofs/src/plonk/permutation/keygen.rs
@@ -9,7 +9,7 @@ use crate::{
     plonk::{Any, Column, Error},
     poly::{
         commitment::{Blind, Params},
-        EvaluationDomain,
+        EvaluationDomain, ProvingKeyTwiddles,
     },
 };
 
@@ -157,6 +157,7 @@ impl Assembly {
         params: &Params<C>,
         domain: &EvaluationDomain<C::Scalar>,
         p: &Argument,
+        fft_twiddles: &ProvingKeyTwiddles<C::Scalar>,
     ) -> ProvingKey<C> {
         // Compute [omega^0, omega^1, ..., omega^{params.n - 1}]
         let mut omega_powers = Vec::with_capacity(params.n as usize);
@@ -197,7 +198,8 @@ impl Assembly {
 
             permutations.push(permutation_poly);
         }
-        let (polys, cosets) = domain.batch_lagrange_to_coeff_and_extended(&permutations);
+        let (polys, cosets) =
+            domain.batch_lagrange_to_coeff_and_extended(&permutations, fft_twiddles);
         ProvingKey {
             permutations,
             polys,